feat:列表服务增加必要参数空指针判断

+++ module-management-service/module-service-list/src/main/java/com/zn/gmall/list/controller/api/ListApiController.java
@@ -4,6 +4,7 @@ import com.zn.gmall.common.result.Result;
 import com.zn.gmall.list.service.api.SearchService;
 import com.zn.gmall.model.list.SearchParam;
 import com.zn.gmall.model.list.vo.SearchResponseVo;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
@@ -15,6 +16,7 @@ import java.io.IOException;
  */
 @RestController
 @RequestMapping("/api/list")
+@Slf4j
 public class ListApiController {
 
     @Autowired
@@ -28,26 +30,37 @@ public class ListApiController {
      */
     @PostMapping("/do/search")
     public Result<SearchResponseVo> list(@RequestBody SearchParam searchParam) throws Throwable {
+        log.info("搜索参数:{}", searchParam);
         SearchResponseVo response = searchService.search(searchParam);
         return Result.ok(response);
     }
 
     @GetMapping("/inner/incr/goods/hot/score/{skuId}")
     public Result<Void> incrGoodsHotScore(@PathVariable("skuId") Long skuId) {
-
+        log.info("商品热度评分增加:{}", skuId);
+        if (skuId == null) {
+            return Result.<Void>fail().message("skuid为空");
+        }
         searchService.incrHotScore(skuId);
-
         return Result.ok();
     }
 
     @GetMapping("/remove/goods/from/elastic/search/{skuId}")
     public Result<Void> removeGoodsFromElasticSearch(@PathVariable("skuId") Long skuId) {
+        log.info("商品从es中删除:{}", skuId);
+        if (skuId == null) {
+            return Result.<Void>fail().message("skuid为空");
+        }
         searchService.removeGoodsFromElasticSearch(skuId);
         return Result.ok();
     }
 
     @GetMapping("/inner/import/sku/to/elastic/search/{skuId}")
     public Result<Void> importSkuToElasticSearch(@PathVariable("skuId") Long skuId) {
+        log.info("商品导入es:{}", skuId);
+        if (skuId == null) {
+            return Result.<Void>fail().message("skuid为空");
+        }
         searchService.importGoodsToElasticSearch(skuId);
         return Result.ok();
     }
